third_party/blink/web_tests/fast/forms/text/text-field-sizing-caret.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
input {
  border: none;
  padding: 0px;
  field-sizing: content;
}
input::-webkit-search-cancel-button {
  display: none;
}
input::-webkit-inner-spin-button {
  display: none;
}
</style>
<input id="text">
<input id="search" type="search">
<input id="number" type="number">

<input id="text-block" style="caret-shape: block">
<input id="search-block" type="search" style="caret-shape: block">
<input id="number-block" type="number"style="caret-shape: block">

<input id="text-underscore" style="caret-shape: underscore">
<input id="search-underscore" type="search" style="caret-shape: underscore">
<input id="number-underscore" type="number" style="caret-shape: underscore">

<script>
// caret-shape: bar
test(() => {
  assert_equals(document.querySelector('#text').offsetWidth, 1);
}, 'The width of an empty text input should be same as the caret width');

test(() => {
  assert_equals(document.querySelector('#search').offsetWidth, 1);
}, 'The width of an empty search input should be same as the caret width');

test(() => {
  assert_equals(document.querySelector('#number').offsetWidth, 1);
}, 'The width of an empty number input should be same as the caret width');

// caret-shape: block
test(() => {
  assert_approx_equals(document.querySelector('#text-block').offsetWidth, 8, 1);
}, 'The width of an empty text input should be same as the width of 1ch for care-shape: block');

test(() => {
  assert_approx_equals(document.querySelector('#search-block').offsetWidth, 8, 1);
}, 'The width of an empty search input should be same as the width of 1ch for care-shape: block');

test(() => {
  assert_approx_equals(document.querySelector('#number-block').offsetWidth, 8, 1);
}, 'The width of an empty number input should be same as the width of 1ch for care-shape: block');

// caret-shape: underscore
test(() => {
  assert_approx_equals(document.querySelector('#text-underscore').offsetWidth, 8, 1);
}, 'The width of an empty text input should be same as the width of 1ch for care-shape: underscore');

test(() => {
  assert_approx_equals(document.querySelector('#search-underscore').offsetWidth, 8, 1);
}, 'The width of an empty search input should be same as the width of 1ch for care-shape: underscore');

test(() => {
  assert_approx_equals(document.querySelector('#number-underscore').offsetWidth, 8, 1);
}, 'The width of an empty number input should be same as the width of 1ch for care-shape: underscore');
</script>
